requirements.txt
telethon
nest_asyncio
python-decouple==3.8
tqdm